refactor(security): remove legacy startup encryption scaffolding
