+++ .env.example
@@ -47,13 +47,6 @@ REDIS_URL=redis://localhost:6379
 # Required when Redis connection URLs are persisted in the database.
 # Generate with: openssl rand -hex 32
 DURABULL_REDIS_URL_ENCRYPTION_KEY=replace-with-64-char-hex-key
-# Temporary migration control. Defaults to true and runs one-time startup migration
-# that encrypts any plaintext redis_connection.url rows on boot.
-# Set to false only after migration cleanup code is removed.
-# DURABULL_REDIS_URL_STARTUP_MIGRATION=true
-# Optional strict mode override used after startup migration is disabled/removed.
-# Defaults to true in production and false in development.
-# DURABULL_ENFORCE_REDIS_URL_ENCRYPTION=true
 
 # Stateless Persistence Mode:
 # - Leave DATABASE_URL unset to use local PGlite persistence at ./data/pglite
