+++ apps/api/src/app.ts
@@ -1,10 +1,4 @@
-import {
-  assertRedisConnectionUrlEncryptionReady,
-  getDatabaseMode,
-  getDb,
-  shouldUseEnvConnections,
-  user,
-} from '@durabull/dal'
+import { getDatabaseMode, getDb, shouldUseEnvConnections, user } from '@durabull/dal'
 import { env } from '@durabull/env'
 import { eq } from 'drizzle-orm'
 import { Hono } from 'hono'
@@ -188,9 +182,6 @@ const getDefaultCorsOrigins = () => {
 
 export async function createApiApp(options: CreateApiAppOptions = {}) {
   const { enableLogging = true, corsOrigins = getDefaultCorsOrigins() } = options
-
-  const db = await getDb()
-  await assertRedisConnectionUrlEncryptionReady(db)
 
   const app = new Hono()
