+++ apps/docs/content/documentation/getting-started/environment-variables.mdx
@@ -31,8 +31,6 @@ description: Complete runtime environment variable reference for local and produ
 | `DATABASE_URL` | Optional | Enables PostgreSQL mode when set |
 | `REDIS_URL` | Optional | Generic Redis URL used by scripts/defaults |
 | `DURABULL_REDIS_URL_ENCRYPTION_KEY` | Required when persisting connections | 32-byte key (hex/base64) used to encrypt Redis connection URLs at rest |
-| `DURABULL_REDIS_URL_STARTUP_MIGRATION` | Temporary | Runs one-time startup migration to encrypt legacy plaintext connection URLs (`true` by default) |
-| `DURABULL_ENFORCE_REDIS_URL_ENCRYPTION` | Optional | Strict plaintext enforcement toggle used when startup migration is disabled (`true`/`false`). Defaults to `true` in production |
 | `DURABULL_PGLITE_DIR` | Optional | Override PGlite data directory |
 
 ## Connection Mode
