+++ apps/docs/content/documentation/meta.json
@@ -31,7 +31,6 @@
     "---Reference---",
     "reference/http-api",
     "---Operations---",
-    "operations/redis-url-migration-cleanup",
     "operations/troubleshooting",
     "deployment/render-and-demo"
   ]
