+++ apps/docs/content/documentation/operations/security-and-hardening.mdx
@@ -29,10 +29,8 @@ Operational guidance:
 
 - grant connection management permissions sparingly
 - separate staging and production Redis credentials
-- startup runs a one-time migration by default to encrypt existing plaintext `redis_connection.url` rows
-- for controlled/manual migration, run `bun tooling/scripts/encrypt-redis-connection-urls.ts`
+- run `bun tooling/scripts/encrypt-redis-connection-urls.ts` before rollout if any legacy plaintext rows exist
 - startup fails fast if encrypted rows cannot be decrypted with the configured key
-- after rollout is complete, follow the cleanup checklist in [Redis URL Migration Cleanup](/documentation/operations/redis-url-migration-cleanup)
 
 ## API Protection
